requirements_test_pre_commit.txt
# From https://github.com/home-assistant/core/blob/dev/requirements_test_pre_commit.txt

bandit==1.6.2
black==19.10b0
codespell==1.16.0
flake8-docstrings==1.5.0
flake8==3.8.1
isort==4.3.21
pydocstyle==5.0.2
pyupgrade==2.3.0
yamllint==1.23.0
